pyproject.toml
[project]
name = "mmu-rag-starter"
version = "0.1.0"
description = "Gemini API integration for TTD-RAG competition"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-genai>=1.0.0",
    "fastapi[standard]>=0.118.0",
    "sse-starlette>=1.6.5",
    "uvicorn>=0.37.0",
    "python-dotenv>=1.1.1",
    "loguru>=0.7.3",
    "pydantic>=2.0",
]
